+++ prueba-infraestructure/src/main/java/pe/gob/pj/prueba/infraestructure/db/negocio/repositories/masters/MaeMateriaRepository.java
@@ -3,6 +3,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import pe.gob.pj.prueba.infraestructure.db.negocio.entities.masters.MaeMateriaEntity;
 import java.util.List;
 
-public interface MaeMateriaRepository extends JpaRepository<MaeMateriaEntity, Integer> {
+public interface MaeMateriaRepository extends JpaRepository<MaeMateriaEntity, Long> {
     List<MaeMateriaEntity> findByActivo(String activo);
 }
