+++ prueba-infraestructure/src/main/java/pe/gob/pj/prueba/infraestructure/db/negocio/repositories/masters/MaeTipoParticipanteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import pe.gob.pj.prueba.infraestructure.db.negocio.entities.masters.MaeTipoParticipanteEntity;
 import java.util.List;
 
-public interface MaeTipoParticipanteRepository extends JpaRepository<MaeTipoParticipanteEntity, Integer> {
+public interface MaeTipoParticipanteRepository extends JpaRepository<MaeTipoParticipanteEntity, Long> {
     List<MaeTipoParticipanteEntity> findByActivo(String activo);
 }
